refactor: split services into shift, skill, and worker modules
